Update PercentileTest.java
Added additional test methods to verify percentile calculation for each of estimation techniques. This is done as part of https://issues.apache.org/jira/browse/MATH-1120. With this; ExcelPercentileTest can be removed as it is subsumed now inside the main Percentile class as an estimation technique (its called R7 technique as per wikiperdia)