Delete ExcelPercentileTest.java
This class will be deleted due to the fact that the Excel style implementation and few others are actually handled inside the Percentile class as separate strategies. Please take a look at Percentile on how excel style is implemented. Suitable tests are also written and is subsumed inside PercentileTest.java.